requirements.txt
# KFit 项目依赖

# 基础依赖
python-dotenv>=0.19.0
PyYAML>=6.0
requests>=2.28.0

# Notion API
notion-client>=1.0.0

# Garmin Connect
garminconnect>=0.1.49

# 大模型API
openai>=1.0.0
anthropic>=0.8.0

# 工具库
python-dateutil>=2.8.2
tqdm>=4.64.0